Modularity
Corrected storage layout to account for multiple modules using the same header
files and exception vectors. See also comments below.

1. Added isprog() function to check if we are in a program block.

2. g instruction now removes files and exception vectors from total space in all
modules but program.

3. When generating address in assemble(), only relocate globals above the
exception vectors per module. This means all files and exceptions are mapped to
the same memory at start of globals.

Now you may note that the exception vectors are divided into public (visible)
and private exceptions used only in pint and pmach. Private vectors means they
are used to handle interpreter errors, and can't be caught by the program.

These are not relocated, they continue the public vectors, and in fact they are
overlapped by the globals from module(s). So what gives?

In the present incantation of the interpreter, exception vectors are single byte
locations, and they don't actually store anything. They have to be unique
addresses, and thus occupy minimum space (byte), but they don't store anything.
They are just used for their unique address. Thus being overlapped by globals
in the program is not a problem. The private exception vectors are only used in
context, so there is no issue.

Of course the truth is all exception vectors can be overlapped with other
globals. However, this would create space allocation issues in pcom, so I didn't
do that. The space to be recovered is minimal.