Modularity
1. Reject option sets in used/joined modules.

2. Only output unreferenced object warnings on procedures/functions. This is
because in Pascaline rules, any module can have global exports. This needs to
also key off the iso7185 flag.

3. Created usesjoins() routine and moved uses/joins handling there.

4. Added save and restore of list and nammod (name of module) for uses/joins.

So we have nearly implemented the same scheme for uses as IP Pascal. Pascaline
does not specify the details of uses/joins. The IP method is that all uses
references are "flat", that is, the used modules are dumped into the same symbol
namespace as everything else, and you get an error if there is a duplicate.

Having said that, this model works, and IP Pascal has used it for 25 years.
However, I don't like it. The main problem is with "zipper effect", which is the
fact that you will pull in definitions from submodules, and submodules of those
modules, etc. This is silent, but can be annoying.

We will see how joins addresses this issue, but I have been working on alternate
plans.