Modularity
1. p6.bat script changed to take any number of files, compile each and
concatenate as a group and run. This allows a run of programs built from
modules.

2. Fixed bad error range definitions in pcom.

3. During includes, having prcode off was generating errors due to bad
stack calculations. I flagged these error checks off when prcode is off.
The other way would have been to make sure that prcode only stops prints, not
things like stack tracking, but that code is far to engrained. It would have
been a complete overhaul for little benefit.

4. Added entname to objects saved before processing includes. Doing it this
way is ugly, but can stay for a while (perhaps forever).