Modularity
1. Changed p6.bat to take any number of input files, compile each one, and
concatenate the intermediates into one and run it.

2. Copied p6 and p6.bat to gpc directory.

3. Moved filptr definitions up.

4. Added external and containing model name to both variable and procedure/
function entries.

5. Added these fields to the table dump routine.

6. Added chkext() routine to check if the identifier structure given is
external.

7. Added prtflabel (print far label) that prints a coined mod.sym name from a
given identifier entry.

8. Modified gencupent() to generate far procedure/function calls.

9. Fixed up gencupent() calls to either pass the called routine or nil, which
signifies a module (since that has no control structure).

10. Created putinp() routine to recycle include stack tracking entries. These
entries are held throughout the run (by necessity). However, it creates errors
to not return entries by the end of the run, and so it was easier to just return
them to free storage.

So what are near vs. far labels? As discussed before, within a module P6
(and previous) use a simple numbered label system and a label table during
assembly. These labels are flushed at the end of each module, and can only
reference locations within that module.

Thus we supplement them by a "far label" system. The difference is that where
near labels have the form:

mod.1

Or a number in the final part, far labels have the form:

mod.sym

Thus a far label can be used to lookup a symbol in any module, and thus cross
module boundaries. Thus a "far" label (not a "far" label like x86 segmented
instructions... don't get me started).