Modularity
Added far labels to external global references.

1. Added symptr to attr, to carry module.symbol parameters.

2. Created gen1s() and gen1ts() versions of gen1() and gen1t() routines that
take a symbol pointer and generate a far label on externals.

3. load(), loadaddress() and store() now use the far version generators along
with the symptr in attr. Each of load and loadaddress also clear the symptr to
break association with the symbol.

4. Load(), loadaddress() and store() were placed together, since they belong
together.

5. Selector() places symptr from the parent symbol in gattr and thus associates
the reference with the symbol. This is only done for variables.

How it works: All variables are gated via load(), loadaddress() and store(). The
only thing that could be called an exception to that is record offsets, which
accumulate the offsets and then get loaded in any case.

P6 (and Pascal-P) use the attr (and gattr) system to track operands through the
code. gattr gets loaded with the essential parameters of a variable reference
when it is loaded or prepared for a store by selector(), and then gattr is
progressively changed as operations are performed on it, either to modify it,
or completely replace its contents.

In old P6 code, gattr does not care about the symbol used to form the reference,
only if it is local or global, what is its address, its type, etc. Now we change
that so that it tracks the original symbol loaded with symptr, but ONLY as long
as needed, ie., until it gets loaded. This then can be used to change the normal
numeric only address into a far label reference. This actually happens in
suprisingly few places.