Modularity
Implemented far labels in pint.

1. Added far label tracking entries.

2. Added far label tracking stack.

3. Changed parlab() to parse either near or far labels. Returns a string with
the far label for far, just the number for near.

4. Added flabrlc() relocation routine for far labels. It clears out the far
label list and executes the patches in it.

5. In generate(), give dummy variable for parlab(), since it will never parse
a far label.

6. Changed labelsearch() to create and enter a far label to the list if found,
otherwise create a near label entry as before.

7. Changed the global load/store instruction handler to either parse a far label
or immediate value. Far labels are introduced by 'l'. Immediate values are fixed
addresses that don't use the labeling system.

This ends the first stage of modularity, and thus multiple modules can be
compiled, run together, and are linked to each other.