Modularity
Implemented the "private" statement.

1. Added priv flag to include file control structures.

2. Force eof on during priv.

3. Force eoln on if eof is on.

4. Add inpriv() function to determine if a private section is active.

5. Suppress errors in includes. Note that pcom's job is not to find errors in
included modules. The information could well be a digest, not a real module.
The way to find errors is to directly compile modules, and the lesson is to
check submodules for errors first.

6. Normalized the appearance of forwarded headers, which is to use:

stuff(x); forward;

stuff{(x)};

This means that procedures can be found by "x(|;|{" (regular expression).

7. Parse and enable private section in declare(). Any use of "private" deeper
than module generates and error, and privates are ignored unless an include is
active. Private sections are then flagged for the rest of the file.

8. In several places we skip parses and supress errors on active private section
to get out of parsing the file after the private section is flagged.

9. Private was added to parse sets.