Modularity
Pint now can handle source level debugging on multiple modules.

1. Changed pcom.pas and pint.pas to have the same name as their modules. This is
a requirement for Pascaline.

2. Line tracking arrays where moved into the block tracking structure.

3. Added bestart field to blocks. This tracks the total span of code in a
module. bstart just tracks the entry point for the module, which can be
different.

4. Added fname to block, which keeps the extended filename for the block.

4. Added curmod, a pointer to the currently active block, which can be nil. It
is determined by pc location, and if the pc is in the pint generated stub there
will be no module active.

5. Added padded length check routine lenp().

6. Added strassfv(), assign variable length to fixed length string.

7. All references to lintrk and linprf arrays are now fully qualified and check
if curmod is non-nil.

8. Block start in generate creates lintrk and linprf arrays in module blocks and
inits them. We don't do this for blocks other than modules, since only modules
have a source file associated with them and it would be a waste of space.

9. Block start creates a full filename for each block and stores in fname.

10. Fndblk was moved, and a new version created fndmod, which only searches for
modules.

11. setcur() procedure was added, it finds the currently executing module via
the pc, or nil if none is found. I attempted to make that fast, especially if
there is no change in block.

12. All references to lintrk and linprf are preceeded by setcur. In the
instruction mrkl, this may have to be revisited, since that is in the execution
path and is time critical.

13. prtsrc was refactored to use curmod and its filename.

14. Source line prints were refactored to fall back to assembly mode if no
source module is active. In some cases, like b (break source) an error results
if no module is active.

Obviously more work to do.