Modularity
1. At module end in generate, reset curmod to nil. Leaving curmod defined after
the block end was putting garbage from the next block into the line tracking
table for the last module.

2. Modified the definition of "end of code" to be the local end of the active
module.

There is an issue that some of the source line definitions in the intermediate
file are outside of any module. There is nothing to be done for that. The reason
is that when we parse a source file, it starts as anonymous, a consequence of
reading from the prd file. We know what the module is when the header for it
is parsed, but any number of lines could have been defined before that time.
In practice, there are a few lines at the end of the module that also exist
in this "no mans land" between any active module.

In debug mode, this does not cause any real issues, as long as both the
programmer and user are aware of the effect. The rule is modules contain the
lines between the start of the module header and the '.' (period) that
terminates it.