Modularity
Added duplicate symbol errors in includes.

1. enterid() checks for in include on duplicate symbols and outputs separate
error messsage if so.

As per previous, errors are suppressed during includes. The policy is that
modules should be compiled as the way to check they are error free, and not
checked as includes. When a "used" module contains duplicates with the current
module, they will get flagged as part of normal processing of the current
module. However, multiple used modules can conflict with each other and that
needs to be flagged. We output a stand-alone error message for these, since
there is no source printout for includes.

Note that joined modules can't duplicate per sey, but a joined module can also
have uses that conflict.