Modularity
First part of support for "joins".

1. Moved the definition of a display record to its own definition.

2. Created the "pile" (tm). See below.

3. Create a new field in displays for module name. It gets the name of the
current module, and that is disposed on block release.

4. Initialized some fields of display that were not touched. Probably not
necessary, but good programming practice.

5. usesjoins() remembers its module type, and creates a copy on the pile for
joined modules. Then the current display record is reinitialized.

6. P6 now enforces the ordering of joins first, uses next. See below.

7. Cleared unused variable out of pmach. There are probably others.

Here is your short theory of how joins works. Joins modules are always placed
onto a clean display (1) (no other definitions). This is a requirement, because
the joined module could be allowed to depend on other modules, a Pascaline
no-no[1].

Joins modules can "use" other modules, and can join them as well, to any depth.
The result will be merged symbols in the display (1).

When a join is complete, it is copied onto the "pile". The pile is a pile
because, unlike the display, there is no particular meaning to its order. The
symbols in various levels of the pile do not nest.

At the end of the joins statement, the pile will contain all of the joined
modules, which may or may not be individually merged with uses, and may also
link to each other with qualidents.

When a symbol is resolved using searchidnenm(), if no symbol is found, it will
turn to the pile and look through the modules there. It will then process
a qualident.

The pile retains the definitions for joined modules until the end of the run,
where they are recycled.

[1] After 10 years of working with Pascaline, I'll admit full well that some of
the rules of Pascaline exist only in my fevered little brain. One of the
purposes of the "rationale" was to dump such rules that would just litter up
the official Pascaline definition.