First commit of pmach.c translation
1. Minor documentation error.

2. In pint/rbf, moved the eof check to within the loop so that all bytes are
checked for eof. There are probably a lot more errors like that.

3. In pint/rsb, removed the requirement that the file must not be closed. This
actually made sense in ISO7185, but not in Pascaline, because the file can
preexist. I did a rules check and didn't find this requirement, but that may
need to be tied to a ISO7185 flag.

4. in pmach, removed begincode declaration. This system is no longer used.

5. In pmach:reads(), created local variable c.

6. First check-in of pmach.c, which is a translation of pmach.pas. This is
preliminary, and does not compile nor work.