p6.bat
1. Adding an extra line to files for mach modes was causing the output binaries
to be different, which complicates debugging. Fixed to output the option for
binary output deck on same line. Note this does not seem to be the only
difference between binary output and regular mode.

2. Changed to use cmach on the option.

cmach.c

1. Added dmpmem routine (normally unused) to aid debugging.

2. All of the references to float were wrong, thats 32 bits. Changed to double,
which matches GPC.

3. Added missing "break"s to case statements.

4. Changed "=" to "==", common error.

5. Changed wrong use of % to /.

6. Removed extra break statement.

7. Changed incorrect == to !=.

pint.pas

1. Added dmpmem() routine for debugging. Note that there is already such a
routine for diagnostic mode, this is just for common debug uses.

2. Changed lstins() use of address to integer, this was causing issues on
certain listings.

pmach.pas

1. Added dmpmem() routine for debugging.