Refactored the scripts completely
1. Scripts now evaluate their parameters regardless of position.

2. Scripts use GNU style --pmach, --cmach, --full, etc. type options. Placement
is not directly relivant, but cannot be last position, see next.

3. Support multi-module linking. Any number of files are accepted and
concatenated into the final intermediate. The last parameter is considered to be
the target program.

4. cmach is now supported.

5. regress has a summary mode, all of the diffs are recapitulated at the end of
the run.

A general comment here: the original use of the compile, run, and p6 scripts was
to be as simple wrappers for each implementation. They are not that anymore. We
may need to get back to that at some point, perhaps by having them call a new,
lower level set of scripts.