regress.bat
Simple comment change.

iso7185prt????.pas

Added {$s+} or ISO 7185 mode switch to files that need it, namely tests that
should fail because of non-standard constructs but don't because Pascaline
allows it. I choose this over adding the flag to all files because the tests
where ISO 7185 and Pascaline don't aligh are a minor subset.