Removed "Pascal-P5 complies with..." message from output compares.
P6 does output this message, but only if the option to enforce ISO 7185 is
build-in, not selected by option. That is because the output of the message
is done before any source is read. I don't feel the need to fix this, so it
stays the way it is and the message is removed from compares.