pint.pas, pmach.pas, cmach.c
1. On donorecpar, add a padding word below the tags if there are zero tags
present in a stack tags record. This is possible on a variant record without
any variant specified in new()/dispose().

2. When disposing with variant (dsl), we place a dummy atom just below the
record address. This replaces the tags length. This allows previously disposed
checks to work on a variant record.

This is a tricky change. donorecpar means that the block returned will be broken
into two parts, an atom at the bottom or heap unit with only the length itself
present, then the rest kept as a permanently occupied unit. This means that
donorecpar is safer than dochkrpt flag, because the record is never reallocated
and thus accesses to fields are simply stale, not accesses to a new and differnt
data item.

If the tags count is zero, that next block breakage will overwrite the tags
count (zero) just under the record base pointer. We avoid that by padding the
tags list at the bottom, giving a place for the block breakage.