Defining point fix
This is a complex fix, see description below.

qsort.pas

Removed an unused variable.

pcom.pas

1. Added a new fatal error maximum. This is actually an error, 512 is not used.

2. Added alias type to identifier structures.

3. Added display flag to indicate if block has (or can have) defining points.

4. Added init of all fields to ininam. I was tired of seeing uninitialized
variable warnings.

5. Moved putnamlst() up and changed name to putparlst(). The old routine still
exists, this is an error. All existing calls to putnamlst() were changed, and
putnam() was refactored to use it.

6. putnam() checks for aliases before disposing of name strings, and does not
do this for aliases, because aliases use the same string as the original
symbol.

7. putnam() added case to dispose of aliases.

8. Added new error for defining point violations.

9. In enterid(), added error handling for if duplicate is normal error or
defining point violation.

10. in searchsection(), if symbol is an alias, we link through to the original
symbol.

11. In schsecidnenm(), we compensate for encountering aliases. We also got rid
of the goto in this routine (I don't like unneeded gotos in the code, and get
rid of them when possible).

12. In searchid(), if we find a symbol, and it is downblock, and the current
block is defining, we create an alias for it. We also got rid of the need for
a goto in this routine.

13. In searchid(), we added a flag to flag aliases as keep, and changed this on
all calls. This may be obsolete. I will check.

13. Added alias case to table prints.

14. Added level print to table dumps to make it clear what level is being
dumped. Before it was dumping all levels merged together.

15. Suppressed referal error on alias.

16. In pushlvl(), we copy the previous block define flag to the new block. This
can be (and is) overridden.

17. In procdeclaration(), we turn the define flag for the block OFF when the
parameter list is being parsed and turn it back ON when the procedure block is
parsed.

The defining point fix adds error detection for the case of an object whose
name is used before that same object is declared in a given block. This is
technically impossible in a serially compiled program, since at the time a given
object is used, it is not known if it will be later defined in that block, or
reference is to an object in a surrounding block. So we instead flag it as an
error when an object is defined within a block, but has been previously used in
the same block.

The way we do this is by establishing an identifier entry known as an "alias".
All it does is serve as a placeholder for a declaration that is outside the
current block. When encountered in an id search, it simply links to the alias
pointed symbol, and upper layers from the searchid() and enterid() calls don't
normally see it. However, enterid(), if it sees a duplicate symbol, will flag
the error differently depending on if the dup is an alias or a regular symbol.
If a regular symbol, it is a normal double defintion. If an alias, it means
there was a previous reference in the block that created an alias of a downlevel
symbol, and therefore is a defining point violation.

Lastly, we divide programming constructs into block levels that are defining
or not defining levels, that is, if they can create new symbols or not. Record
field lists, parameter lists, and "with" references cannot. A flag was added to
the display that indicates, when the block level is created, if it is a defining
block or not. Recall that block levels are created for lots of reasons,
including to isolate routines, field lists and with references.
The only tricky one here is procedure/functions, which don't define in the
parameter list section (parameters only feature type ids), but do define
elsewhere in the block. For this we modulate the defining flag to partition the
parameters from the rest of the block.

The definining point flag on the block works by supressing alias creation for
that block. This has the side effect of saving space in the compiler, since
alias entries are not needed in non-definining blocks.