pint.pas
Changed the way numbers are calculated for different word widths. The
new way should work for any power of 2 word size, 16, 32, 64, 128, etc.

1. Each power gets a maximum power and a digit count. The binary count,
being the number of bits in a word, has other uses as well.

2. Changed wrtnum to use calculated powers and digit counts.

3. Changed octal calculations for 1 or 2 digits left over. These are
the only possibilities, since 3 would only occur if the word size were
and odd number of bits.

4. Created routine to calculate powers and digits.

5. Init powers and digits.