64 bit adaption
pint.pas

Fixed return statements (retp,retc,retb,reti,retx,retr and reta) to be
correct for both 64 and 32 bit results.

On all models, reals are the biggest elements that are returned on the
stack. In 64 bit, reals and integers are the same size (unextended).
On 32 and 16 bit, they are half the size of real. Thus we must consider
if the basic stack element or stackelsize is the same or less than
maxresult, and correct the return placement accordingly.