1st part of container arrays
pcom.pas

1. Added arrayc structure definition. This is the same as arrays, but has no
index type. This is added to all prints, disposals, etc.

2. Moved stackbot to locals. It didn't need to be global.

3. Added new error for parameterized variables.

4. Added arrayc to qualification in stringt(). Containers are considered strings
if they are packed and have a base type of char. Note they are always indexed
1..n.

5. Moved expression() and all its subcomponents up above declare() and into
global space. This also needed the transfer of checkbnds(), load(), and
loadaddress(). I also moved store(), because it belongs with that group of
routines. This was all necessary because a delcare routine needs to process
expressions, vardeclaration(). This is one of the uglier things about
parameterized types, see the rationale on this. Because these routines were
de-indented, this refactor looks worse than it is.

6. Added a skeleton of parameterized types to vardeclaration(). This just parses
expressions and discards them.

This is the first installment of a fairly long patch. It passes regression.