basics.pas
1. Fixed out of bounds error that occurred during testing. If the
line prtlin() sees is entirely blank, j will go to zero and the
access str[j] will go out of bounds. Refactored.

pcom.pas

1. On the previous fix for 64 bit results, the check in selector()
for where the function result lies was wrong. it needs to check
both if the result is smaller than maxresult and if the minimum
stack size is less than maxresult.