Container arrays, continuing series
pcom.pas

1. Added vis and vip instructions (vector initialize stack and vector
initialize pointer).

2. Added code strip label list system to display.

3. Added new errors for initializers.

4. Extended genujpxjp() to genujpxjpcal(), its just a name change.

5. Added new "v" or vector type to intermediate type entry system.
Even multidemensional arrays are considered vectors of vectors.
Vectors and arrays are no different, but we term variable length arrays
to be vectors.

6. Added container count function containers() to vardeclaration().

7. Added container initializer code strip generation to vardeclaration().
These are called by the code in the block it belongs to.

8. Added calls to initializer code strips to body(). Note that these are
near calls. The label entries are disposed of.

There are quite a few more container array changes to come, this is
a long series.

Just for grins we are 8296 lines in pcom.pas vs. 6090 lines for P5.
Not bad considering how much capability was added.