String library
strings.pas

Added the string library from IP Pascal, released under the BSD
2-clause license (or "FreeBSD" license).

This is the first of several libraries that will come from the
IP Pascal project and be released as open source. It is written
to a previous verion of Pascaline and requires some fixup. It
also needs both containers and overloads, so is aspirational
at the moment. I choose it as the first import library because it
has few dependencies.

This is the first use of the BSD license. The plan is for this
to become the general license for original works in Pascal-p6.