strings library
strings.pas

Cleaned up the code and removed "IP"isms.

Note I am still a bit vague on the exact details of the class
layout with respect to ths module, but realize it is important to
define.

The principle issue is that the two basic string types that the
module deals with, fixed strings and pstrings, can both be extended
as types and methods without placing them within an object. Another
way to say that is that the entire string library is itself in a
static object, the strings module itself. Thus it needs research
on exactly what needs doing.

Thus the module goes forward without classes for now.