pcom.pas
1. Added arrtmp() routine to create fixed templates from normal
array specs. This is needed because arrays are created in different
areas.

2. Added template label to common array type.

3. Output templates for several areas in the code.

pint.pas

1. Added new license spec. See below.

2. Added generate() spec to generate fixed templates.

Fixed templates are used to interface container arrays with fixed
(normal ISO 7185 Pascal arrays). Anytime we see a combination of
fixed and container arrays, the method is we promote the fixed array
to a container by loading a template pointer on top of the base
pointer and thus converting fixed to container array specs.

P6 generates such templates for every array, there is no attempt
to join them together or reduce them. This is not considered a
big space consumer.

After long consideration, I decided we are going heavy into BSD
licensing, which is the fairest and freest license, since it also
allows full commercial use. The merits and drawbacks have been
discussed (endlessly) elsewhere.

Applying the BSD licence to a public domain source like P6 (and
it clearly is PD) means that only the incremental improvements made
by me are BSD. The PD content can be determined by diffs with P4
(the last unadulterated PD version of Pascal-P), or you just use
P4.

I don't see that as a big problem in any case, since BSD is about
the loosest of the licences. In any case, P6 is not a GPL candidate
since it is "tainted" by the base of PD code.