Makefile
Added clean targets.

bin/*

Added new binaries to track and check everything in. The binaries are
tracked so the project will work "out of the box". Note we static link
everything so that the chances of this are good.

cmach.c, genobj.pas, pmach.pas, spew.c

Updated the license information on these, now the license is "BSD for
changes", IE, The BSD license for everything I have added to the
project, not covering the original public domain sources of Pascal-P4.

pcom.pas

1. Added new instructions, equv, neqv, gegv, grtv, leqv and lesv. These
are vector (container array) equivalents of the old string compare
instructions.

2. Moved code from assignment() that rationalizes mixed container/
normal array binary operations into complex pointers only into new
routine containerop(). This is now used in expressions as well.

3. In expression(), string compares now check for containers and issue
new vector operators for this.

4. Pint added new vector instructions equv, neqv, gegv, grtv, leqv
and lesv.