cmach.c, pmach.pas
Changed the operand order of both the wrs and wrsp system routines.
They now take field width, base address, and length of array.

This was done to normalize the appearance of strings with container
arrays.

pcom.pas

1. In containerop(), corrected right case for new complex pointer
layout.

2. In selector() array index operation, replaced complex to simple
convertion from just loading a length, to using the scp (Simplify
Complex Pointer) instruction.

3. In writeprocedure(),, if a simple string is loaded (address only),
then it is converted to complex pointer on load. In this way,
containers and standard strings are unified, making the code
simpler.

4. In writeprocedure(), if no field exists on a container string, we
have to do some stack shuffling because the length is above the
pointer. We have to duplicate that as the field. A standard string
just gets the fixed field default loaded.

pint.pas

1. Changed the operand order of both the wrs and wrsp system routines
as above.

2. Fixed bug in cxc instruction, because we advance over the first
template entry, we need to loop over one less entry.