pcom.pas
1. Changed name of scp (simplify complex pointer) instruction to
spc (simplify pointer complex). scp is now store complex pointer,
and matches lcp (load complex pointer).

2. Added ccs (complex copy to stack) and ldp (load double pointer).

3. Added flag to var entry for "contains double pointer" see below.

4. In loadaddress(), if the double pointer flag is set, the form
is double pointer and we load that. If not, we use lcp and simplify
if the container level is 1. Note that double pointer is a simplified
pair if container level 1.

5. In selector(), copy double pointer flag to attributes.

6. In actual parameter copy, we now split off complex arrays and
use a different method to create value copies. Instead of
preallocating the space on the stack in parameterlist(), we do
that at runtime and copy the data in place with the ccs instruction.
This instruction takes a complex pointer at stack top, then allocates
that copy space on the stack rounded to even stack size, then
replaces the complex pointer on stack with a new pointer to the copy.
This is then stored as the parameter copy.

pint.pas

1. Changed scp to spc as stated above.

2. Added ccs, scp and ldp instructions.

3. Removed stack balance checks on return instructions. Sadly,
since that was a great cross check. However, we now allocate things
on the stack and there is no easy way to account for that. The
stack is simply reset on returns. The error vector for stackbalance
is now unused.

There are several forms (as it turns out) for complex pointers. The
normal allocation for cps is a base address followed by the template
itself. A pointer can be created from that, by taking a pointer to
the template after the base pointer. There is also a simplified
form of that which has the length of a single level container array
instead of the template pointer.

Parameters are special in that they contain either a simplified
lenth or the address of the template as that second pointer, NOT
the template itself. Thus we need to load a double pointer. This
actually works on both simplified and regular pointers.