pcom.pas
1. Added two new instructions, vin (vector initialize new) and vdd
(vector deinitialize dispose). vin is like vip, but for dynamic
allocation, see below. vdd is identical to vdp, it disposes the given
dynamic variable, but is a match for vin.

2. Added spana(), a function to get the span, or index range, of an
array.

3. Replaced calls to span(), for the index range, with spana(). The
difference is that span() does not correctly evaluate the length
of string constants, which use a nil index.

4. Added special load in selector() for dynamic pointers to containers.
See below.

5. Changed new/dispose of containers to use vin and vdd.

pint.pas

1. Added vin and vdd instructions.

new/dispose of container pointers had to be changed to use a special
procedure. The issue is that all pointers have to be the same size
by implication of ISO 7185 rules. The reason is they can be forward
references, and can be built into any type tree while they are
forwarded. This works as long as all pointers are the same size,
normally just an address. Container complex pointers don't even fit
that requirement, since any number of template entries can exist
for one.

The answer is to store the template in the first part of the dynamic
allocation for the container array, then skip over that template to
get the actual data. This means that pointers go back to being
simple addresses. This system is different from the "store behind
the data" model used with tag checking, because container arrays
always need their templates, vs. Tag checking data can be simply
ignored if that checking is not done.