Changes for fixed declarations.
Updated main make file for make help.

Ran regression.

1. Added ltc and lto instructions (load table contant and load table
offset) (ltc in existing slot).

2. load() generates ltc instructiions for fixed table pickups. This
is the same as ldo, but indicates the operand is in constant space.

3. loadaddress() generates lto instruction for fixed table pickups.
This is the same as lao, but indicates the operand is in constant
space.

4. Removed unneeded comment.

1. Added ltc and lto instructions. These are turned back into ldo and
lao instructions, but we enter their addresses into the constants area
fixups.

2. Added table generation to generate().

This completes the basics for fixed declaration. I will do one more to
generate container constant arrays (yes, there is really such a thing).