Changes for fixed declarations.
Regressed.

1. Updated gen1s to output near labels for fixed.

2. constfactor() fixed to accept empty sets.

3. Reversed order for constant table sets. See below.

4. Fixed string outputs on fixed.

1. In generate() constant table build, skip spaces before constant
set.

2. Added lto convertion case on assemble().

When parsing numbers vs. symbols in intermediate text, the existing
code has an issue with mixing numbers and symbols (single character).
It uses a read ahead mechanisim using "ch" as the read ahead. The
issue is that the code also uses normal reads to get numbers, and
the built in reads know nothing about the lookahead. When executing
a read, you must know that either a space or number exits in the
input, and if there is a symbol expected that means something, you
either need to see that or a space. The example here is a set, which
looks like:

( 1 2 3)

The key is the space before each number, which gets eaten by the
lookahead. Thus, the code reads the space, knows that a ')' is not
forthcoming, and the list continues.

If instead the list is:

(1 2 3)

The lookahead would have eaten the '1' character trying to find if
list ends. In the case of a null list, it would in fact end without
giving a number.

I beleive this is an artifact of the P4 code that didn't implement
file buffer variables, and thus had no built-in lookahead. It is no
longer true, but would need the code to be extensively refactored
(see the Pascal-P mission statements in the docs).

The upshot is that the intermediate must be carefully constructed
to avoid problem cases.