Fixes for fixed declarations.
Ran regression.

Selector() was rolling record address offsets into the starting
address, which does not work with externals/constant tables since
they are based on a label. We fix up load() and loadaddress() to
handle the case where the offset is non-zero by generating extra
code for this case.

This is a bit dirty, but the extra code only exists for new
features, so it isn't so bad.

Note that I believe this also fixes record offsets to externals, so
this fixes an existing issue.