Add containers to fixed declarations.
Regress.

1. Create new chained entry for fixed length "initializers". When a fixed
is made from a container, it gets the inits from the data, and stores them
in a chain, since any number of levels is possible.

2. Created getfcl(), putfcl(), disposals for fixed identifier entries,
and tracking of those entries.

3. fixeditem() now takes a var containing the current initializer entry.
For each container array, it is called with an initial nil pointer. The
first subcontainer found sets that, then all the others must match it.
Example:

array of array of ...

Is two levels deep. fixeditem calls itself recursively for each sublevel.
The first one determines its length and creates the entry. All subsequent
calls compare the length and error if no match. This recurses to any
number of levels.

4. Added arrayc (container) handling in fixeditem(), as specified.
Each time a length entry is created, it is pushed onto the fixed identifier
entry as a list.

5. At top level, we check for a list of initializers and reverse it if
found. This changes it from depth first order to major order.